Moderate code cleanup for 9x writing

+++ core/src/main/java/net/vhati/openuhs/core/CRC16.java
@@ -124,7 +124,7 @@ public class CRC16 implements Checksum {
 
 	@Override
 	public void update( int n ) {
-		sum = a[(int) ((sum >> 8) & 0xFF)] ^ ((sum & 0xFF) << 8) ^ b[n & 0xFF];
+		sum = a[((sum >> 8) & 0xFF)] ^ ((sum & 0xFF) << 8) ^ b[n & 0xFF];
 		sum &= 0xFFFF;
 	}
 }
